서버 연동 수정
일단 사이즈를 줄이는 쪽으로 가고 총알 행렬을 분리시키는게 좋을듯

+++ Project Client/PillsFighter/GameFramework.cpp
@@ -852,11 +852,11 @@ void CGameFramework::recvn()
 	char *pBuf = m_RecvBuf;
 	int nPacketsize = 0;
 
-	if (m_nPacketSize > 0) nPacketsize = (int)m_pPacketBuffer[0];
+	if (m_nPacketSize > 0) nPacketsize = m_pPacketBuffer[0];
 
 	while (nRest > 0)
 	{
-		if (nPacketsize == 0) nPacketsize = (int)pBuf[0];
+		if (nPacketsize == 0) nPacketsize = pBuf[0];
 
 		int nRequired = nPacketsize - m_nPacketSize;
